Работа над подзадачей №8
